update  points congrats
