+++ packages/huma-shared/src/utils/chain.ts
@@ -2,7 +2,7 @@ import type { AddEthereumChainParameter } from '@web3-react/types'
 import { ethers } from 'ethers'
 
 export enum CHAIN_TYPE {
-  EVM = 'ethereum',
+  EVM = 'evm',
   SOLANA = 'solana',
 }
