+++ packages/huma-widget/src/components/Lend/solanaSupply/index.tsx
@@ -12,13 +12,13 @@ import { setStep } from '../../../store/widgets.reducers'
 import { selectWidgetState } from '../../../store/widgets.selectors'
 import { WIDGET_STEP } from '../../../store/widgets.store'
 import { ErrorModal } from '../../ErrorModal'
+import { PointsEarned } from '../../PointsEarned'
 import { WidgetWrapper } from '../../WidgetWrapper'
 import { Evaluation } from './1-Evaluation'
 import { ChooseTranche } from './2-ChooseTranche'
 import { ChooseAmount } from './3-ChooseAmount'
 import { Transfer } from './4-Transfer'
 import { Success } from './5-Success'
-import { PointsEarned } from './6-PointsEarned'
 
 export interface Campaign {
   id: string
@@ -59,7 +59,6 @@ export function SolanaLendSupply({
   const [tokenAccount, isLoadingTokenAccount] = useTokenAccount(poolInfo)
   const { step, errorMessage } = useAppSelector(selectWidgetState)
   const [selectedTranche, setSelectedTranche] = useState<TrancheType>()
-  const [transactionHash, setTransactionHash] = useState<string | undefined>()
 
   useEffect(() => {
     if (!step && !isLoadingLenderAccounts && !isLoadingTokenAccount) {
@@ -141,24 +140,23 @@ export function SolanaLendSupply({
         />
       )}
       {step === WIDGET_STEP.Transfer && selectedTranche && (
-        <Transfer poolInfo={poolInfo} selectedTranche={selectedTranche} />
+        <Transfer
+          poolInfo={poolInfo}
+          poolState={poolState}
+          selectedTranche={selectedTranche}
+          pointsTestnetExperience={pointsTestnetExperience}
+        />
       )}
       {step === WIDGET_STEP.Done && (
         <Success
           poolInfo={poolInfo}
           poolState={poolState}
           campaign={poolState.campaign}
-          updateTransactionHash={setTransactionHash}
           handleAction={handleClose}
         />
       )}
-      {step === WIDGET_STEP.PointsEarned && transactionHash && (
-        <PointsEarned
-          transactionHash={transactionHash}
-          poolState={poolState}
-          pointsTestnetExperience={pointsTestnetExperience}
-          handleAction={handleClose}
-        />
+      {step === WIDGET_STEP.PointsEarned && (
+        <PointsEarned lpConfig={poolState} handleAction={handleClose} />
       )}
       {step === WIDGET_STEP.Error && (
         <ErrorModal
