+++ packages/huma-widget/src/components/Lend/supplyV2/5-Transfer.tsx
@@ -1,11 +1,17 @@
-import { PoolInfoV2, TrancheType } from '@huma-finance/shared'
+import {
+  CampaignService,
+  checkIsDev,
+  PoolInfoV2,
+  TrancheType,
+} from '@huma-finance/shared'
 import { useTrancheVaultContractV2 } from '@huma-finance/web-shared'
 import { useWeb3React } from '@web3-react/core'
 import { ethers } from 'ethers'
 import React, { useCallback } from 'react'
 
+import { Campaign } from '.'
 import { useAppDispatch, useAppSelector } from '../../../hooks/useRedux'
-import { setStep } from '../../../store/widgets.reducers'
+import { setPointsAccumulated, setStep } from '../../../store/widgets.reducers'
 import { selectWidgetState } from '../../../store/widgets.selectors'
 import { WIDGET_STEP } from '../../../store/widgets.store'
 import { TxSendModalV2 } from '../../TxSendModalV2'
@@ -13,14 +19,19 @@ import { TxSendModalV2 } from '../../TxSendModalV2'
 type Props = {
   poolInfo: PoolInfoV2
   trancheType: TrancheType
+  pointsTestnetExperience: boolean
+  campaign?: Campaign
 }
 
 export function Transfer({
   poolInfo,
   trancheType,
+  pointsTestnetExperience,
+  campaign,
 }: Props): React.ReactElement | null {
+  const isDev = checkIsDev()
   const dispatch = useAppDispatch()
-  const { account, provider } = useWeb3React()
+  const { account, chainId, provider } = useWeb3React()
   const { supplyAmount } = useAppSelector(selectWidgetState)
   const { decimals } = poolInfo.poolUnderlyingToken
   const supplyBigNumber = ethers.utils.parseUnits(
@@ -34,9 +45,26 @@ export function Transfer({
     account,
   )
 
-  const handleSuccess = useCallback(() => {
-    dispatch(setStep(WIDGET_STEP.Done))
-  }, [dispatch])
+  const handleSuccess = useCallback(
+    async (options?: { txHash: string }) => {
+      if (campaign && options?.txHash) {
+        try {
+          const result = await CampaignService.updateWalletPoints(
+            account!,
+            options.txHash,
+            chainId!,
+            isDev,
+            pointsTestnetExperience,
+          )
+          dispatch(setPointsAccumulated(result.pointsAccumulated))
+        } catch (error) {
+          console.error('Failed to update wallet points', error)
+        }
+      }
+      dispatch(setStep(WIDGET_STEP.Done))
+    },
+    [account, campaign, chainId, dispatch, isDev, pointsTestnetExperience],
+  )
 
   if (!trancheVaultContract || !account) {
     return null
