+++ packages/huma-widget/src/store/widgets.store.ts
@@ -59,6 +59,7 @@ export type WidgetState = {
   stream?: WidgetStream
   multisend?: MultisendPayload
   solanaSignature?: string
+  pointsAccumulated?: number
 }
 
 export const initialWidgetState: WidgetState = {}
